Added an initial contrl unit
Untested, though it compiles. :)

PUSH and POPMULTI need a few more muxes; there are ten now. So
this may change. But all the planned instructions are implemented
including a load/store indirect with an integrated 8 bit
displacement, which will be useful for accessing the stack.

Renamed left to reg2 and right to reg3, since there are 3 registers
extracted from the instruction word now. reg1 is generally the
destination so does not enter the ALU.