Added the outter entity, including the datapath
All fairly streightforward. Changed the type T_ALU_OPS to be 5 bits wide,
taking the MSB from the LSB of the opcode, in the same way as the 16 bit
core did. Whilst I could make the "quick" byte portion 7 bits and still
have a useable set of one word ALU operations, it would mean the register
indirect with quick displacement would either use a different width
quick immediate or it would similarly loose a bit.

Anyway, much of this will likely be changed around when the testbench
for the whole processor is started, which will be next.