Switch to GHDL
ModelSim was previously used, but GHDL is much nicer for various
reasons, including being FOSS.

The main change is that GHDL does not seem to like circular
dependancies. Previously control dep'd on P_REGS and registers
depended on P_CONTROL. Because the packages were included in the
same file as the implementation a circular dependancy was created
Move the packages, containing only the externally used definitions,
to their own VHDL file, and compile these first, to avoid the
problem. Private elements, like the OPCODE constants, have been
moved inside the entity in question. The exception to this is
currently the ALU ops, which need to be a package so the test
bench can use them.

Also renamed P_REGS to P_REGISTERS to attempt to be more consistent.

The Makefile itself remains rubbish since it does not use
depenancy tracking to avoid rebuilds etc. Need to investigate how
to write proper Makefiles for GHDL projects.