Added an initial customasm CPU definition
The CPU definition is missing multi stack ops, but has pretty much
everything else including short hand versions of the ALU ops, so
you can do:

add r0,r1

Which means the same as:

add r0,r0,r1

Ie. it assumes the first operand and the destinationa are the same,
otherwise we'd have code filled with weird things like:

not r0,r0

Instead of:

not r0

Also added an initial go at a BRAM VHDL implementation for the
CPU test bench. The RAM is 256 x 32bit, with byte selects which
are honoured on writes but not reads (since that doesn't make
much sense).