Running a test program in simulation!
Albeit one that does nothing but load a register and increment it in a
loop. :)

The biggst change needed was to rejig the ALU inputs, which was done
in the previous commit.

Needs some Makefile work to make it so the test programs can be rebuilt
and then fed into the hex2package.pl script.

Lots and lots of cleanups to do.