Rework CALL*
The return address was being written into the stack using the address
of the subroutine instead of the address after it, so the return
calls were broken. Reworked all the flow opcodes to be in one
when block. This includes return, which can now use the condition
codes, though the cpu defintion hasn't yet been extended to make
use of this. Not sure how useful conditional returns are...