Further refinements
Made embedded quick values sign extended.

Fixed problems with RDQ operations and added them to the fib test bench.
Currently this quick displacement is a byte offset, but it could/should
be a long offset instead to give us two more bits, since the stack is always
long aligned.

Renamed the quick bytes and words "quick" and not "imm".

Added all the operations to the README.md file and updatd the wording.