Switch the conditions to an ARM style 4 bit code
Instead of using 8 bits for the condition, where most of the 256 available
values make no sense at all, use a ARM-style 4 bit code to cover all
of the useful ones.

There is one difference: the codes are shifted down one, so 0 (and F)
are the "always" case.

This leaves 12 bits free in the instruction longword, which will be handy
for a quick branch instruction, which will be done next.