Added a strings asm module for playing with the ISA
This will eventually turn into a machinecode monitor, but for now this is
just to validate the ISA design. So far so good. For comparison the 68K
high level instructiosn of:

    move.b %d0,(%a0)+
    bne foo

ends up as:

    load.bu r0,(r1)
    addq r1,#1
    test r0
    branchqne foo

The lack of postincrement is totally reasonable.  Though interstingly the
address index is fed into the indcdec_index in the register file (as used by
stacking), and there is a bit free in the needed instruction format.  There
would be other work to do however, especially because increments of 1, 2 and
4 would need to be added.

Not manipulating the condition flags after a load is unusual and I've not
seen a CISC design which behaves like this.  But wihtout also doing post
increments without involving the ALU, the CFs would end up being adjusted by
the add operation.  Unless a flag was added to ALU instructions which
enabled/disabled the CF changing, though there are no free bits on the
ALU quick format, there are plenty of spare opcodes in those "pages".

As it stands, I'm pretty happy with things.