Some progress with running code on the FPGA, and code for a UART
To make write memory operations work with the BRAM, the processor clock was
further slowed, relative to the external (memory) clock.  Somewhat annoying.

uart.asm is the starting point for a monitor.  Read and write of
stringsseems to be working, using a simple UART I cooked up on MAXI030 as a
base.  The UART hardware itself is on the level above this repo.

Also fixed some logic in the conditional tests, since ARM reverses the sense
of the carry flag on subtractions.