Redo register to register copies in a dedicated opcode
The previous method, using the ALU with a mechanism for supressing
the flag change, wasn't right. Added a new opcode (and prefix) for
register to register copies. This required a new mux input on the
register input mux, for the 2nd register output, which completes the
loop.

Left the old ALU-based copy in place for now.

In the future this prefix could be extended with condition cade
flag copying, etc.